+++ package/app/lib/ui/component/common/app_logo_widget.dart
