+++ package/app/lib/ui/page/top_page.dart
@@ -9,7 +9,7 @@ import '../component/common/navigation_button/to_gallery_button.dart';
 import '../component/common/navigation_button/to_quiz_setting_button.dart';
 import '../component/play_quiz_daily/start_todays_daily_quiz_button.dart';
 import '../component/play_quiz_normal/start_normal_quiz_from_top_button.dart';
-import '../component/setting/app_logo_widget.dart';
+import '../component/common/app_logo_widget.dart';
 import '../component/setting/open_setting_button.dart';
 import '../controller/top_page_controller.dart';
